adding restricion page

+++ studybud_clone/base/views.py
@@ -1,29 +1,35 @@
-from django.shortcuts import render,redirect
+from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
-from .models import Room,Topic
+from .models import Room, Topic
 from .forms import RoomForm
 from django.contrib.auth.models import User
-from django.contrib.auth import authenticate, login,logout
+from django.contrib.auth import authenticate, login, logout
 # Create your views here.
+
 
 def home(request):
     q = request.GET.get("q") if request.GET.get("q") != None else ''
     rooms = Room.objects.filter(
         Q(topic__name__icontains=q) |
-        Q(name__icontains=q) | Q(description__icontains=q)  
+        Q(name__icontains=q) | Q(description__icontains=q)
     )
     topics = Topic.objects.all()
-    context = {"rooms": rooms,"topics":topics}
+    context = {"rooms": rooms, "topics": topics}
     return render(request, "base/home.html", context)
 
 
-def room(request,pk):
+def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {"room":room}
-    return render(request, "base/room.html",context)
+    context = {"room": room}
+    return render(request, "base/room.html", context)
 
+
+@login_required(login_url="login")
 def createRoom(request):
+    print("this is the request",request.user)
     form = RoomForm()
     if request.method == "POST":
         # print(request.POST)
@@ -31,28 +37,39 @@ def createRoom(request):
         if form.is_valid():
             form.save()
             return redirect('home')
-    context = {"form":form}
-    return render(request, "base/room_form.html",context)
+    context = {"form": form}
+    return render(request, "base/room_form.html", context)
 
-def updateRoom(request,pk):
+
+@login_required(login_url="login")
+def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
+    if request.user != room.host :
+        return HttpResponse("You are not allowed here!")
+
     if request.method == "POST":
-        form = RoomForm(request.POST,instance=room)
+        form = RoomForm(request.POST, instance=room)
         if form.is_valid():
             form.save()
             return redirect('home')
-        
-    context = {"form":form}
-    return render(request,'base/room_form.html',context)
 
-def deleteRoom(request,pk):
+    context = {"form": form}
+    return render(request, 'base/room_form.html', context)
+
+@login_required(login_url="login")
+def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host :
+        return HttpResponse("You are not allowed here!")
+    
     if request.method == "POST":
         room.delete()
         return redirect("home")
-    return render(request,"base/delete.html",{"obj":room})
+    return render(request, "base/delete.html", {"obj": room})
+
 
 def loginPage(request):
     if request.method == "POST":
@@ -62,17 +79,18 @@ def loginPage(request):
         try:
             user = User.objects.get(username=username)
         except:
-            messages.error(request,"User does not exist")
+            messages.error(request, "User does not exist")
 
-        user = authenticate(request,username=username,password=password)
+        user = authenticate(request, username=username, password=password)
 
         if user is not None:
-            login(request,user)
+            login(request, user)
             return redirect("home")
         else:
-            messages.error(request,"Username or Password is incorrect")
+            messages.error(request, "Username or Password is incorrect")
     context = {}
-    return render(request, "base/login_register.html",context)
+    return render(request, "base/login_register.html", context)
+
 
 def logoutUser(request):
     logout(request)
